+++ test/foundry/Raffle.enterRaffles.t.sol
@@ -38,7 +38,7 @@ contract Raffle_EnterRaffles_Test is TestHelpers {
         assertEq(user2.balance, 0.975 ether);
         assertEq(address(looksRareRaffle).balance, 0.025 ether);
 
-        (uint256 amountPaid, uint256 entriesCount, bool refunded) = looksRareRaffle.rafflesParticipantsStats(1, user2);
+        (uint208 amountPaid, uint256 entriesCount, bool refunded) = looksRareRaffle.rafflesParticipantsStats(1, user2);
 
         assertEq(amountPaid, 0.025 ether);
         assertEq(entriesCount, 1);
@@ -70,7 +70,7 @@ contract Raffle_EnterRaffles_Test is TestHelpers {
         assertEq(user2.balance, 0);
         assertEq(address(looksRareRaffle).balance, 1.17 ether);
 
-        (uint256 amountPaid, uint256 entriesCount, bool refunded) = looksRareRaffle.rafflesParticipantsStats(1, user2);
+        (uint208 amountPaid, uint256 entriesCount, bool refunded) = looksRareRaffle.rafflesParticipantsStats(1, user2);
 
         assertEq(amountPaid, 1.17 ether);
         assertEq(entriesCount, 110);
@@ -95,7 +95,7 @@ contract Raffle_EnterRaffles_Test is TestHelpers {
         assertEq(user2.balance, extra);
         assertEq(address(looksRareRaffle).balance, price);
 
-        (uint256 amountPaid, uint256 entriesCount, bool refunded) = looksRareRaffle.rafflesParticipantsStats(1, user2);
+        (uint208 amountPaid, uint256 entriesCount, bool refunded) = looksRareRaffle.rafflesParticipantsStats(1, user2);
 
         assertEq(amountPaid, price);
         assertEq(entriesCount, 1);
